o  use both environ and __environ as available 	o  forward decl fixes for copy_regexs() and copy_funcs()

+++ glish/Func.h
@@ -189,6 +189,6 @@ glish_declare(PList,UserFunc);
 typedef PList(UserFunc) func_list;
 
 extern void describe_parameter_list( parameter_list* params, OStream& s );
-extern void copy_funcs( void *to_, void *from_, unsigned int len );
+extern void copy_funcs( void *to_, void *from_, size_t len );
 
 #endif /* func_h */
